requirements.txt
Flask~=3.0.3
flask-restx~=1.3.0
google-cloud-language~=2.13.4